2038 year problem

+++ src/helpers/datetimehelper.h
@@ -1,8 +1,9 @@
 #ifndef DATETIMEHELPER_H
 #define DATETIMEHELPER_H
 
+#include <inttypes.h>
 
-static int getYears(int time)
+static int getYears(int64_t time)
 {
     return time/(60*60*24*365);
 }
